Avoid stale keys, defaultValue, etc
When the key changes dynamically, it still uses the old key since methods are not depending properly on the key. With this PR, when the key changes the methods change accordingly. It also avoid stale default values. TBH, the whole dependency was quite troublesome and I had to clean up a bunch of code.